feat(reachability_check): track the result of all the dial attempts
